+++ NEWS.md
@@ -1,3 +1,8 @@
+# tinyVAST 0.3.0
+
+* Adding `sdmTMB` as dependency, and importing family options from it
+* Adding vignette for joint analysis of condition and density
+
 # tinyVAST 0.2.0
 
 * Add option to specify covariance in SEM and DSEM notation